requirements.txt
pytz
django
google-cloud-spanner==1.2.0
